🔨 Refactor(#28): 파일 위치 수정
