+++ src/features/part/types.ts
@@ -1,6 +1,5 @@
-interface Part {
+export interface Part {
   id: number;
   sectionId: number;
   name: string;
 }
-export default Part;
