+++ src/features/part/ui/PartForm.tsx
@@ -1,38 +1,37 @@
-import { FloatingLabel, Form } from 'react-bootstrap';
-
-import usePartStore from '../../../store/usePartStore';
+import { Button, FloatingLabel, Form } from 'react-bootstrap';
 import sectionsQueries from '../../../queries/sections';
+import getFormDataValue from '../service/getFormData';
 
 export default function PartForm() {
   const { data: sections } = sectionsQueries.read();
-  const { part, pushPart } = usePartStore();
+  const handleMutate = (formData: FormData) => {
+    const sectionId = getFormDataValue(formData, 'sectionId');
+    const part = getFormDataValue(formData, 'part');
+
+    console.log(sectionId, part);
+  };
   return (
     <>
-      <Form.Select
-        id="sectionId"
-        className="mb-3"
-        onChange={e => {
-          pushPart(e.target.id as 'sectionId', Number(e.target.value));
-        }}
-      >
-        <option>섹션 선택</option>
-        {sections?.map(section => (
-          <option key={section.id} value={section.id}>
-            {section.name}
-          </option>
-        ))}
-      </Form.Select>
-      <FloatingLabel label="파트 입력">
-        <Form.Control
-          id="name"
-          size="sm"
-          type="text"
-          defaultValue={part.name}
-          onChange={e => {
-            pushPart(e.target.id as 'name', e.target.value);
-          }}
-        />
-      </FloatingLabel>
+      <Form action={handleMutate}>
+        <Form.Select id="sectionId" className="mb-3" name="sectionId">
+          <option>섹션 선택</option>
+          {sections?.map(section => (
+            <option key={section.id} value={section.id}>
+              {section.name}
+            </option>
+          ))}
+        </Form.Select>
+        <FloatingLabel label="파트 입력">
+          <Form.Control
+            id="part"
+            size="sm"
+            type="text"
+            name="part"
+            onChange={e => {}}
+          />
+        </FloatingLabel>
+        <Button type="submit">asdasds</Button>
+      </Form>
     </>
   );
 }
